+++ src/lib/Scenes/Articles/Articles.tsx
@@ -2,15 +2,11 @@ import { Articles_articlesConnection$key } from "__generated__/Articles_articles
 import { ArticlesQuery } from "__generated__/ArticlesQuery.graphql"
 import { LoadFailureView } from "lib/Components/LoadFailureView"
 import { extractNodes } from "lib/utils/extractNodes"
-import { PlaceholderBox, ProvidePlaceholderContext, RandomWidthPlaceholderText } from "lib/utils/placeholders"
-import _ from "lodash"
-import { Flex, Separator, Spacer } from "palette"
 import React, { useEffect, useState } from "react"
-import { FlatList } from "react-native"
 import { ConnectionConfig } from "react-relay"
 import { usePagination, useQuery } from "relay-hooks"
 import { graphql } from "relay-runtime"
-import { ArticlesHeader, ArticlesList, ArticlesListItem, useNumColumns } from "./ArticlesList"
+import { ArticlesList, ArticlesPlaceholder } from "./ArticlesList"
 
 const PAGE_SIZE = 10
 
@@ -69,38 +65,6 @@ export const ArticlesQueryRenderer: React.FC = () => {
   }
 
   return <ArticlesPlaceholder />
-}
-
-export const ArticlesPlaceholder: React.FC = () => {
-  const numColumns = useNumColumns()
-
-  return (
-    <ProvidePlaceholderContext>
-      <Flex flexDirection="column" justifyContent="space-between" height="100%" pb={8}>
-        <Separator />
-        <FlatList
-          numColumns={numColumns}
-          key={`${numColumns}`}
-          ListHeaderComponent={() => <ArticlesHeader />}
-          data={_.times(6)}
-          keyExtractor={(item) => `${item}-${numColumns}`}
-          renderItem={({ item }) => {
-            return (
-              <ArticlesListItem index={item} key={item}>
-                <PlaceholderBox aspectRatio={1.33} width="100%" marginBottom={10} />
-                <RandomWidthPlaceholderText minWidth={50} maxWidth={100} marginTop={1} />
-                <RandomWidthPlaceholderText height={18} minWidth={200} maxWidth={200} marginTop={1} />
-                <RandomWidthPlaceholderText minWidth={100} maxWidth={100} marginTop={1} />
-                <Spacer mb={2} />
-              </ArticlesListItem>
-            )
-          }}
-          ItemSeparatorComponent={() => <Spacer mt="3" />}
-          onEndReachedThreshold={1}
-        />
-      </Flex>
-    </ProvidePlaceholderContext>
-  )
 }
 
 const fragmentSpec = graphql`
