+++ src/lib/Scenes/Articles/ArticlesList.tsx
@@ -1,14 +1,15 @@
 import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
 import { ArticleCard_article } from "__generated__/ArticleCard_article.graphql"
 import { ArticleCardContainer } from "lib/Components/ArticleCard"
+import { PlaceholderBox, ProvidePlaceholderContext, RandomWidthPlaceholderText } from "lib/utils/placeholders"
 import { ProvideScreenTrackingWithCohesionSchema } from "lib/utils/track"
 import { screen } from "lib/utils/track/helpers"
 import { useScreenDimensions } from "lib/utils/useScreenDimensions"
+import _ from "lodash"
 import { Flex, Separator, Spacer, Text } from "palette"
 import React from "react"
 import { ActivityIndicator, FlatList, RefreshControl } from "react-native"
 import { useTracking } from "react-tracking"
-
 interface ArticlesListProps {
   articles: ArticleCard_article[]
   isLoading: () => boolean
@@ -16,7 +17,7 @@ interface ArticlesListProps {
   refreshing: boolean
   handleLoadMore: () => void
   handleRefresh: () => void
-  title: any
+  title: string
 }
 
 export const ArticlesList: React.FC<ArticlesListProps> = ({
@@ -112,6 +113,38 @@ export const useNumColumns = () => {
   return orientation === "portrait" ? 2 : 3
 }
 
+export const ArticlesPlaceholder = () => {
+  const numColumns = 1
+
+  return (
+    <ProvidePlaceholderContext>
+      <Flex flexDirection="column" justifyContent="space-between" height="100%" pb={8}>
+        <Separator />
+        <FlatList
+          numColumns={numColumns}
+          key={`${numColumns}`}
+          ListHeaderComponent={() => <ArticlesHeader />}
+          data={_.times(6)}
+          keyExtractor={(item) => `${item}-${numColumns}`}
+          renderItem={({ item }) => {
+            return (
+              <ArticlesListItem index={item} key={item}>
+                <PlaceholderBox aspectRatio={1.33} width="100%" marginBottom={10} />
+                <RandomWidthPlaceholderText minWidth={50} maxWidth={100} marginTop={1} />
+                <RandomWidthPlaceholderText height={18} minWidth={200} maxWidth={200} marginTop={1} />
+                <RandomWidthPlaceholderText minWidth={100} maxWidth={100} marginTop={1} />
+                <Spacer mb={2} />
+              </ArticlesListItem>
+            )
+          }}
+          ItemSeparatorComponent={() => <Spacer mt="3" />}
+          onEndReachedThreshold={1}
+        />
+      </Flex>
+    </ProvidePlaceholderContext>
+  )
+}
+
 export const tracks = {
   tapArticlesListItem: (articleId: string, articleSlug: string) => ({
     action: ActionType.tappedArticleGroup,
@@ -123,7 +156,7 @@ export const tracks = {
   }),
 }
 
-export const ArticlesHeader = ({ title = "Market News" }) => (
+export const ArticlesHeader = ({ title = "" }) => (
   <Text mx="2" variant="lg" mb={1} mt={6}>
     {title}
   </Text>
